requirements.txt
beautifulsoup4==4.12.2
pandas>=1.5.0
requests==2.31.0
urllib3==2.1.0
lxml==4.9.3
streamlit>=1.24.0
plotly>=5.13.0
schedule==1.2.1 
chardet
psutil
sqlalchemy>=1.4.0
python-dateutil
google-generativeai>=0.3.0


fastapi
uvicorn
python-dotenv>=0.19.0
PyYAML 
openpyxl